# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "REST API for phpMyFAQ 4.1",
    "description": "phpMyFAQ includes a REST API and offers APIs for various services like fetching the phpMyFAQ version or doing a search against the phpMyFAQ installation.",
    "contact": {
      "name": "phpMyFAQ Team",
      "email": "support@phpmyfaq.de"
    },
    "license": {
      "name": "Mozilla Public Licence 2.0",
      "url": "https://www.mozilla.org/MPL/2.0/"
    },
    "version": "3.1"
  },
  "servers": [
    {
      "url": "https://localhost",
      "description": "Local dockerized server"
    }
  ],
  "paths": {
    "/api/v3.1/attachments/{faqId}": {
      "get": {
        "tags": ["Public Endpoints"],
        "description": "Returns a list of attachments for a given FAQ record ID.",
        "operationId": "getAttachments",
        "parameters": [
          {
            "name": "faqId",
            "in": "path",
            "description": "The FAQ record ID.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If the FAQ has at least one attached file.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the attachment.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "\n        [\n            {\n                \"filename\": \"attachment-1.pdf\",\n                \"url\": \"https://www.example.org/index.php?action=attachment&amp;id=1\"\n            },\n            {\n                \"filename\": \"attachment-2.pdf\",\n                \"url\": \"https://www.example.org/index.php?action=attachment&amp;id=2\"\n            }\n        ]"
              }
            }
          },
          "404": {
            "description": "If the FAQ has no attachments.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the attachment.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": []
              }
            }
          }
        }
      }
    },
...
    "/api/v3.1/categories": {
      "get": {
        "tags": ["Public Endpoints"],
        "operationId": "getCategories",
        "responses": {
          "200": {
            "description": "Returns the the categories for the given language provided by \"Accept-Language\".",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the categories.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "\n        [\n            {\n                \"id\": 1,\n                \"lang\": \"en\",\n                \"parent_id\": 0,\n                \"name\": \"Test\",\n                \"description\": \"Hello, World! Hello, Tests!\",\n                \"user_id\": 1,\n                \"group_id\": 1,\n                \"active\": 1,\n                \"show_home\": 1,\n                \"image\": \"category-1-en.png\",\n                \"level\": 1\n              }\n        ]"
              }
            }
          },
          "404": {
            "description": "If no categories are found for the given language.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the categories.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": []
              }
            }
          }
        }
      }
    },
    "/api/v3.1/category": {
      "post": {
        "tags": ["Endpoints with Authentication"],
        "operationId": "createCategory",
        "requestBody": {
          "description": "The parent category ID is a required value, the parent category name is optional. If the parent category name is present and the ID can be mapped, the parent category ID from the name will be used. If the parent category name cannot be mapped, a 409 error is thrown",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "language",
                  "parent-id",
                  "parent-category-name",
                  "category-name",
                  "description",
                  "user-id",
                  "group-id",
                  "is-active",
                  "show-on-homepage"
                ],
                "properties": {
                  "language": {
                    "type": "string"
                  },
                  "parent-id": {
                    "type": "integer"
                  },
                  "parent-category-name": {
                    "type": "string"
                  },
                  "category-name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "user-id": {
                    "type": "integer"
                  },
                  "group-id": {
                    "type": "integer"
                  },
                  "is-active": {
                    "type": "boolean"
                  },
                  "show-on-homepage": {
                    "type": "boolean"
                  }
                },
                "type": "object"
              },
              "example": "{\n                \"language\": \"en\",\n                \"parent-id\": 1,\n                \"parent-category-name\": \"Test\",\n                \"category-name\": \"Test 2\",\n                \"description\": \"Hello, World! Hello, Tests!\",\n                \"user-id\": 1,\n                \"group-id\": 1,\n                \"is-active\": true,\n                \"show-on-homepage\": true\n            }"
            }
          }
        },
        "responses": {
          "201": {
            "description": "If all posted data is correct.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the login.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"stored\": true }"
              }
            }
          },
          "400": {
            "description": "If something didn't worked out.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the login.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"stored\": false, \"error\": \"Cannot add category\" }"
              }
            }
          },
          "409": {
            "description": "If the parent category name cannot be mapped.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the login.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {},
                "example": "{ \"stored\": false, \"error\": \"The given parent category name was not found.\" }"
              }
            }
          },
          "401": {
            "description": "If the user is not authenticated.",
            "headers": {
              "Accept-Language": {
                "description": "The language code for the login.",
                "schema": {
                  "type": "string"
                }
              },
              "x-pmf-token": {
                "description": "phpMyFAQ client API Token, generated in admin backend",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}